Fix selectors matching against arrays containing subdocuments.
This is a 0.5.5 regression (with the new selector compiler): the new compiler
handled selectors with "a.b.c" reasonably well if the "c" level contained an
array but not if the "a" or "b" levels did.

Could use some more tests (of more than one level of depth, of other selectors,
etc).

Should also use this in sorts (though that is not believed to be a 0.5.5
regression).